Configure clippy and lint on push (#52)
Enable as many lints as I can stand. I'll definitely add more exceptions as
`clippy::pedantic` and `clippy::restriction` wear me down.

+++ src/info.rs
@@ -4,7 +4,7 @@ use crate::common::*;
 pub struct Info {
   pub private: u8,
   #[serde(rename = "piece length")]
-  pub piece_length: u64,
+  pub piece_length: u32,
   pub name: String,
   #[serde(with = "serde_bytes")]
   pub pieces: Vec<u8>,
